Add Aspose.Cells Cloud 25.8 new feature to Aspose.Cells Cloud SDK.

+++ lib/aspose_cells_cloud/requests/convert_range_to_image_request.rb
@@ -168,7 +168,9 @@ module AsposeCellsCloud
       form_params = {}
       post_body = nil 
       header_params['Content-Type'] = api_client.select_header_content_type(['multipart/form-data'])
-      form_params[File.basename(spreadsheet)] =  ::File.open(spreadsheet,"r")   
+      if !spreadsheet.empty? && File.exist?(spreadsheet )
+          form_params[File.basename(spreadsheet)] =  ::File.open(spreadsheet,"r")   
+      end 
 
       #auth_names = []
       auth_names = ['JWT']
